+++ examples/aws_driver_example/aws_failover_postgresql_example.ts
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 
-import { AwsPGClient } from "../../pg";
+import { AwsPgClient } from "../../pg";
 import { FailoverFailedError, FailoverSuccessError, TransactionResolutionUnknownError } from "../../index";
 
 const postgresHost = "db-identifier.XYZ.us-east-2.rds.amazonaws.com";
@@ -23,7 +23,7 @@ const password = "employees";
 const database = "database";
 const port = 5432;
 
-const client = new AwsPGClient({
+const client = new AwsPgClient({
   // Configure connection parameters, failover plugin enabled by default.
   host: postgresHost,
   port: port,
@@ -71,12 +71,12 @@ try {
   await client.end();
 }
 
-async function setInitialSessionSettings(client: AwsPGClient) {
+async function setInitialSessionSettings(client: AwsPgClient) {
   // User can edit settings.
   await client.query("SET TIME ZONE UTC");
 }
 
-async function queryWithFailoverHandling(client: AwsPGClient, query: string) {
+async function queryWithFailoverHandling(client: AwsPgClient, query: string) {
   try {
     const result = await client.query(query);
     return result;