+++ examples/aws_driver_example/telemetry_tracing_xray_example.ts
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 
-import { AwsPGClient } from "../../pg";
+import { AwsPgClient } from "../../pg";
 import { AsyncHooksContextManager } from "@opentelemetry/context-async-hooks";
 import { context } from "@opentelemetry/api";
 import { NodeSDK } from "@opentelemetry/sdk-node";
@@ -65,7 +65,7 @@ process.on("SIGTERM", () => {
     .finally(() => process.exit(0));
 });
 
-const client = new AwsPGClient({
+const client = new AwsPgClient({
   user: "username",
   host: "db-identifier.XYZ.us-east-2.rds.amazonaws.com",
   database: "database_name",